requirements.txt
django
tensorflow==1.15
ckiptagger[tf,gdown]
beautifulsoup4
opencc-python-reimplemented
